# docs/04-api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NextCRM API",
    "description": "NextCRM is a comprehensive Customer Relationship Management system with multi-tenant architecture, subscription billing, and project management capabilities. This API provides programmatic access to all CRM features including accounts, contacts, leads, opportunities, invoices, projects, and document management.",
    "version": "3.0.0-alpha",
    "contact": {
      "name": "NextCRM Support",
      "url": "https://github.com/DrivenIdeaLab/nextcrm-app",
      "email": "support@nextcrm.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.nextcrm.io/api",
      "description": "Production API Server"
    },
    {
      "url": "https://demo.nextcrm.io/api",
      "description": "Demo API Server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/projects/tasks/update-task/{taskId}": {
      "put": {
        "summary": "Update a task",
        "description": "Update task details",
        "tags": ["Projects"],
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/user/{userId}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve user information",
        "tags": ["Users"],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from NextAuth authentication"
      }
    },
    "schemas": {
...
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "priority": {
            "type": "string"
          },
          "dueDateAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "string"
          },
          "taskStatus": {
            "type": "string",
            "enum": ["ACTIVE", "PENDING", "COMPLETE"]
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "userStatus": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "PENDING"]
          },
          "userLanguage": {
            "type": "string",
            "enum": ["en", "de", "cz", "uk"]
          },
          "is_admin": {
            "type": "boolean"
          },
          "organizationId": {
            "type": "string",
            "nullable": true
          },
          "organization_role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "MEMBER", "VIEWER"],
            "nullable": true
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_on": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error", "message"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or session invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthenticated",
              "message": "Valid authentication token is required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource was not found or user does not have access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not Found",
              "message": "Account not found or unauthorized",
              "code": "NOT_FOUND"
            }
          }
        }
      },
...
    }
  }
}